Rename game source folder from gameboard to iq_game
