+++ src/iq_game/components/board_segment.py
