+++ src/iq_game/digits/zero.py
@@ -1,6 +1,6 @@
-from gameboard.digits import Digit
-from gameboard.utils import grid_utils
-from gameboard.components import BoardSegment, BoardSquare
+from iq_game.digits import Digit
+from iq_game.utils import grid_utils
+from iq_game.components import BoardSegment, BoardSquare
 
 
 class Zero(Digit):
