+++ src/iq_game/game_board.py
@@ -1,9 +1,9 @@
 from typing import List, Tuple
 from copy import deepcopy
 
-from gameboard.components import BoardSegment, BoardSquare
-from gameboard.utils import grid_utils
-from gameboard.digits import Digit
+from iq_game.components import BoardSegment, BoardSquare
+from iq_game.utils import grid_utils
+from iq_game.digits import Digit
 
 
 class BoardGrid:
