+++ tests/test_place_crossovers/test_one_crossing_seven.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.append('./src/')
 
-from gameboard import BoardGrid
+from iq_game import BoardGrid
 
 
 def test_place_one_over_seven_up(setup_game_state: dict):
