+++ tests/test_place_crossovers/test_seven_crossing_one.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.append('./src/')
 
-from gameboard import BoardGrid
+from iq_game import BoardGrid
 
 
 def test_place_seven_over_one_up(setup_game_state: dict):
