+++ tests/test_place_individual_digits/test_place_two.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.append('./src/')
 
-from gameboard import BoardGrid
+from iq_game import BoardGrid
 
 
 def test_place_two_up(setup_game_state: dict):
